camera_test: Queue tap-out directly to tap-in
Add support to queue tap-out buffer directly to tap-in surface. This
is only possible if the stride of the tap-out buffer is aligned
properly for ISS. We will fall back to previous method of copying
the tap-out buffer to tap-in if this condition is not met.

Depends on system/core change Ic65652a5:
http://review.omapzoom.org/#/c/24869/

Depends on frameworks/native change Ia5f695b3:
http://review.omapzoom.org/#/c/25509/

Change-Id: I186bd7e7a41074cb7fcfe1961b709b96e5187b2c
Signed-off-by: Tyler Luu <tluu@ti.com>
Signed-off-by: Vladimir Petrov <vppetrov@mm-sol.com>
Signed-off-by: Daniel Levin <dendy@ti.com>