Camera_test: Fix typo in absolute exposure settings
There is a typo in setSingleExpGainPreset().
When in absolute exp. bracketing modes, the address of the
abs() math function is used as value to configure the exposure
time, instead of the exposure time argument.
After applying this patch, the absolute exposure time
in streaming shot mode will be set correctly.

Change-Id: I244dd6869f2b69bf5d7c087c694001e7b64d032d
Signed-off-by: Vladimir Petrov <vppetrov@mm-sol.com>