Camera_test: Video mode fix
When start recoding camera is not in video mode but in HQ mode.
This is because the width and height are different for preview and video recording.
Preview is restarted and in this case the mode is lost and the default HQ mode is set.
To prevent this behaviour width and height for the preview are set the same as for video recording.

Change-Id: I72df65f7732b3f4067cd4b3e9b5047081a0b0dd7
Signed-off-by: Dimitar Borisov <x0152684@ti.com>