camera_test: Fix for disconnect and reconnect to camera
[Problem]
Currently when try to disconnect and reconnect to
camera using "@" command, segmentation fault is seen.
This is happening because when open camera, some of the parameters
are not set to valid valiues when setParameters function is called.

[Solution]
When get parameters form capabilities, initDefaults() function is called
to set all the parameters to the default values and then setParameters is called.

Change-Id: Icd49ca6286b92badec9c302c966263139aa204f3
Signed-off-by: Dimitar Borisov <x0152684@ti.com>