hwc: Avoid blits when cloning to TV
The workaround to avoid blits when playing a video was not allowing
the blitter to be used when using only the LCD (single display).

It was found by the display team the stuttering issue with video
playback is not GC320 but rather a bug in dsscomp to manage correctly
the synchronization when using multiple displays.

This patch revisits the workaround by not allowing to blit when
multiple displays are used but allowing all blits with a single
display use case.

This allows not only to prevent the stuttering when playing a video
(with the TV connected), it also allows any application that refreshes
the screen constantly to not stutter (like a 3D game) by using the
GPU to compose instead of the blitter.

The reason why the GPU performs better with the TV connected is the
compositor gets blocked on the availability of the flip chain buffers,
thus preventing very short periods of time between posts caused by the
VSYNC drifting of both displays not being handled correctly by dsscomp.
With this change a more consistent framerate is achieved when cloning
to TV.

Change-Id: Iedc93c3528095fcf24f4a34b8ac13744ece2b576
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>