hwc: Do not swap red and blue channels if we can blit all
There is a special case when the composition can be handled by
the DSS if all layers have the HAL_PIXEL_FORMAT_RGBA_8888 by
swapping the red and blue channels in the manager.

If we are using BLTPOLICY_ALL and the blitter can handle all
layers we don't need to swap the red and blue channels. This
was leading to see incorrect colors when we can blit all and all
layers have the HAL_PIXEL_FORMAT_RGBA_8888 format. The issue can
be fixed by avoiding the swap when using the blitter and when
BLTPOLICY_ALL is used.

This issue doesn't happen with BLTPOLICY_DEFAULT.

Change-Id: Id426b239b7cc99131aa426a7f3300b167dd135e3
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>