hwc: Fix for External Display legacy mode mirroring
During legacy mode mirroring, we have to bypass
all the layers from Surfaceflinger to avoid any GLES
composition for the external display. This was
achieved by marking all the layers for HWC_OVERLAY
composition type.
The patch removes the check for valid layer to mark
the layer for HWC_OVERLAY, because it was skipping some
of the layers which do not meet mem1d and scaling requirements
from the HW.

Change-Id: I558a7462a72594014943ed3e66fef419fb0ad4af
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>