CameraHal: Avoid to call OMX_Deinit() twice
In the destructor of OMXCameraAdapter, OMX_Deinit() is called after OMX state
is switched to Loaded via switchToLoaded function.
But if some error happens during the state change, OMX_Deinit() is called in
switchToLoaded function.
As the result, OMX_Deinit() is called twice in swtchToLoaded and the destructor.
This causes the ducati crash.

This patch adds checking for the mOmxInitialized after switchToLoaded() to
avoid double deinitialization.

Change-Id: Ie30170de2bf032a1861a00b2ca370266fe12d339
Signed-off-by: Nanako Baba <baba@ti.com>
Signed-off-by: Daniel Levin <dendy@ti.com>