hwc: rgz: Workaround, do not blit if there is a video layer
This is a workaround for a performance degradation found while
a video is playing and the GC320 is used at the same time.
Avoid blitting if there is a NV12 layer in the hwc layer list.

Reapplication of the I12e789b9dd95e8b0d435aac1ae67ff65d4511396
patch due to regression still being present.

Change-Id: I09a3010dd55ddf4b65db1967c42c4ea3640cf4f9
Signed-off-by: Pavlo Suikov <pavlo.suikov@ti.com>